34 | Add like button with fontawesome and animate.css

+++ articles/views.py
@@ -1,10 +1,11 @@
 from IPython import embed
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden, HttpResponse
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_POST, require_GET
-# from accounts.models import User
 from django.contrib.auth import get_user_model
+
 from .forms import ArticleForm, CommentForm
 from .models import Article, Comment
 
@@ -60,13 +61,14 @@ def detail(request, article_pk):
     return render(request, 'articles/detail.html', context)
 
 @require_POST
+@login_required
 def delete(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    # if request.method == 'POST':
-    article.delete()
-    return redirect('articles:index')
-    # else:
-    #     return redirect('articles:detail', article.pk)
+    if article.user == request.user:
+        article.delete()
+        return redirect('articles:index')
+    else:
+        return HttpResponseForbidden()
 
 # def edit(request, article_pk):
 #     article = Article.objects.get(pk=article_pk)
@@ -74,39 +76,44 @@ def delete(request, article_pk):
 #         'article': article
 #     }
 #     return render(request, 'articles/edit.html', context)
-
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        article_form = ArticleForm(request.POST, instance=article)
-        if article_form.is_valid():
-            article = article_form.save()
-            return redirect('articles:detail', article.pk)
+    if article.user == request.user:
+        if request.method == 'POST':
+            article_form = ArticleForm(request.POST, instance=article)
+            if article_form.is_valid():
+                article = article_form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            article_form = ArticleForm(instance=article)
+        context = {
+            'article_form': article_form
+        }
+        return render(request, 'articles/form.html', context)
     else:
-        article_form = ArticleForm(instance=article)
-    context = {
-        'article_form': article_form
-    }
-    return render(request, 'articles/form.html', context)
+        return HttpResponseForbidden()
 
 @require_POST
 def comment_create(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    # 1. modelform에 사용자 입력값 넣고
-    comment_form = CommentForm(request.POST)
-    # 2. 검증하고,
-    if comment_form.is_valid():
-    # 3. 맞으면 저장!
-        # 3-1. 사용자 입력값으로 comment instance 생성 (저장은 X)
-        comment = comment_form.save(commit=False)
-        # 3-2. FK 넣고 저장
-        comment.article = article
-        comment.save()
-    # 4. return redirect
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        # 1. modelform에 사용자 입력값 넣고
+        comment_form = CommentForm(request.POST)
+        # 2. 검증하고,
+        if comment_form.is_valid():
+        # 3. 맞으면 저장!
+            # 3-1. 사용자 입력값으로 comment instance 생성 (저장은 X)
+            comment = comment_form.save(commit=False)
+            # 3-2. FK 넣고 저장
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+        # 4. return redirect
+        else:
+            messages.success(request, '댓글이 형식이 맞지 않습니다.')
+        return redirect('articles:detail', article_pk)
     else:
-        messages.success(request, '댓글이 형식이 맞지 않습니다.')
-    return redirect('articles:detail', article_pk)
-
+        return HttpResponse('Unauthorized', status=401)
     # comment = Comment()
     # comment.content = request.POST.get('content')
     # comment.article = article
@@ -116,15 +123,19 @@ def comment_create(request, article_pk):
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    # messages.add_message(request, messages.INFO, '댓글이 삭제 되었습니다.')
-    messages.success(request, '댓글이 삭제되었습니다.')
-    return redirect('articles:detail', article_pk)
+    if request.user == comment.user:
+        comment.delete()
+        # messages.add_message(request, messages.INFO, '댓글이 삭제 되었습니다.')
+        messages.success(request, '댓글이 삭제되었습니다.')
+        return redirect('articles:detail', article_pk)
+    else:
+        return HttpResponseForbidden()
 
 @login_required
 def like(request, article_pk):
-    article = Article.objects.get(pk=article_pk)
+    article = get_object_or_404(Article, pk=article_pk)
     # 좋아요를 누른적이 있다면?
+    # if article.like_users.filter(id=request.user.id):
     if request.user in article.like_users.all():
         # 좋아요 취소 로직
         article.like_users.remove(request.user)
@@ -133,4 +144,3 @@ def like(request, article_pk):
         # 좋아요 로직
         article.like_users.add(request.user)
     return redirect('articles:detail', article_pk)
-
